backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
motor==3.3.2
redis==5.0.1
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
pymongo==4.6.1
twilio==8.11.1
slack-sdk==3.26.2
reportlab==4.0.9
python-docx==1.1.0
groq==0.13.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
authlib==1.3.0
itsdangerous==2.1.2
email-validator==2.1.0
